pyproject.toml
[project]
name = "smarttour"
version = "0.1.0"
description = "SmartTour travel planning PoC with FastAPI and Streamlit."
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "folium>=0.18.0",
    "googlemaps>=4.10.0",
    "httpx>=0.28.0",
    "openai>=1.60.0",
    "ortools>=9.11",
    "pydantic-settings>=2.0",
    "sqlmodel>=0.0.24",
    "streamlit>=1.40.0",
    "streamlit-folium>=0.22.0",
    "uvicorn[standard]>=0.34.0",
]

[project.scripts]
seed-database = "scripts.seed_database:main"

[project.optional-dependencies]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.14.6"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["scripts*", "smarttour*"]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
required-environments = [
    "sys_platform == 'win32'"
]

[tool.ruff]
exclude = ["docs"]

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E501"]

[tool.mypy]
disable_error_code = ["import-untyped"]
exclude = ["docs"]
python_version = "3.12"
